# aphrodite-backend/docs/swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "aphrodite",
        "description": "API Documentation",
        "version": "unknown"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8083/aphrodite-web-api",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/v1/key-value/data/{bucket}": {
            "get": {
                "tags": ["KeyValueData"],
                "operationId": "readKeyValueData",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Secret Key Version",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        },
                        "example": 1
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "put": {
                "tags": ["KeyValueData"],
                "operationId": "writeKeyValueData",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Secret Key Version",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        },
                        "example": 1
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "additionalProperties": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/WriteKeyValueDataResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "delete": {
                "tags": ["KeyValueData"],
                "operationId": "deleteKeyValueData",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "Secret Key Version",
                        "required": false,
                        "allowEmptyValue": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        },
                        "example": 1
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
        "/api/v1/login/userpass": {
            "post": {
                "tags": ["Login"],
                "operationId": "userpassLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPassLoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    }
                }
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
        "/api/v1/key-value/metadata/{bucket}": {
            "get": {
                "tags": ["KeyValueMetadata"],
                "operationId": "getKeyValueMetaData",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/KeyValueMetaData"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
        "/api/v1/key-value/metadata/{bucket}/list-keys": {
            "get": {
                "tags": ["KeyValueMetadata"],
                "operationId": "listKeyValueKeys",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Secret Key Path",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": ""
                        },
                        "example": "app/config/db-password"
                    },
                    {
                        "name": "bucket",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Aphrodite-API-Key": []
                    }
                ]
            },
            "parameters": [
                {
                    "name": "AphroditeNamespace",
                    "in": "header",
                    "description": "Namespace header for API requests",
                    "required": false,
                    "allowEmptyValue": true,
                    "schema": {
                        "type": "string",
                        "default": "aphrodite"
                    }
                }
            ]
        },
...
    },
    "components": {
        "schemas": {
...
            "WriteKeyValueDataResponse": {
                "type": "object",
                "properties": {
                    "version": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "UserPassLoginRequest": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "validitySeconds": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
            "LoginResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    }
                }
            },
...
            "KeyValueMetaData": {
                "type": "object",
                "properties": {
                    "lastCreatedAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "firstCreatedAt": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "versions": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Aphrodite-API-Key": {
                "type": "apiKey",
                "name": "Aphrodite-API-Key",
                "in": "header"
            }
        }
    }
}